pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "jinja2>=3.1.6",
    "joblib>=1.5.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "uvicorn>=0.35.0",
    "weasyprint>=65.1",
    "xgboost>=3.0.2",
]
